requirements.txt
colorlog>=4.0.2
python-rapidjson>=0.7.0
pyyaml>=4.2b1
Jinja2>=2.10.1
